refactor: add as_regular_shape helper routine
