+++ tests/conftest.py
@@ -24,6 +24,7 @@ from zarr.core.array import (
 )
 from zarr.core.chunk_grids import (
     SHARDED_INNER_CHUNK_MAX_BYTES,
+    as_regular_shape,
     guess_chunks,
     normalize_chunks_nd,
     resolve_outer_and_inner_chunks,
@@ -378,7 +379,7 @@ def create_array_metadata(
         return ArrayV2Metadata(
             shape=shape_parsed,
             dtype=dtype_parsed,
-            chunks=tuple(int(dim[0]) for dim in outer_chunks),
+            chunks=as_regular_shape(outer_chunks),
             order=order_parsed,
             dimension_separator=chunk_key_encoding_parsed.separator,
             fill_value=fill_value,
@@ -397,7 +398,7 @@ def create_array_metadata(
         sub_codecs: tuple[Codec, ...] = (*array_array, array_bytes, *bytes_bytes)
         codecs_out: tuple[Codec, ...]
         if inner is not None:
-            inner_chunks_flat = tuple(int(dim[0]) for dim in inner.outer_chunks)
+            inner_chunks_flat = as_regular_shape(inner.outer_chunks)
             index_location = None
             if isinstance(shards, dict):
                 index_location = ShardingCodecIndexLocation(shards.get("index_location", None))
